// src/api/generated/models/modusOperandiResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */

export interface ModusOperandiResponse {
  id?: string;
  crimeTypeCode?: string;
  name?: string;
  description?: string;
  warningSigns?: string[];
  recommendations?: string[];
}