// src/api/generated/models/reporterHistoryResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */
import type { PreviousCall } from './previousCall';

export interface ReporterHistoryResponse {
  reporterId?: string;
  contactCount?: number;
  previousCalls?: PreviousCall[];
}